fix(codeql): resolve high/note severity CodeQL findings in inventory module
- Use long arithmetic in getAvailableQuantity(), reservation newQuantity,
  and transfer TRANSFER_IN/TRANSFER_OUT newQuantity to prevent int overflow/underflow
- Wrap Long.parseLong in try-catch to handle NumberFormatException in inventoryMovements resolver
- Remove unused @AuthenticationPrincipal principal params from reserve, release,
  confirm (InventoryController) and cancelTransfer (InventoryTransferController)

+++ src/main/java/ecommerce/modules/inventory/entity/Inventory.java
@@ -98,7 +98,7 @@ public class Inventory {
     }
 
     public int getAvailableQuantity() {
-        return Math.max(0, onHandQuantity - reservedQuantity);
+        return (int) Math.max(0L, (long) onHandQuantity - reservedQuantity);
     }
 
     public InventoryStatus getStatus() {
